Refactored old commands and queries, removed redunant

+++ src/CommunicationsMicroservice/Communications.Application/Reports/Commands/DeleteReport/DeleteReportCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Communications.Core.Interfaces;
+﻿using Communications.Core.CustomExceptions;
+using Communications.Core.Interfaces;
 using MediatR;
 
 namespace Communications.Application.Reports.Commands;
@@ -12,6 +13,11 @@ public class DeleteReportCommandHandler : ReportsHandlerBase, IRequestHandler<De
     public async Task Handle(DeleteReportCommand request, CancellationToken cancellationToken)
     {
         var report = await GetIfExistsAsync(request.Id);
+
+        if (!report.CanBeEdited)
+        {
+            throw new BadRequestException("Report can not be edited anymore.");
+        }
 
         UnitOfWork.ReportsRepository.Delete(report);
